requirements.txt
asgiref==3.7.2
Django==4.2.7
django-extensions==3.2.3
djangorestframework==3.14.0
pytz==2024.1
sqlparse==0.4.4
